image_pipeline_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
